requirements.txt
discord.py==0.16.12
riotwatcher==2.2.1
pandas==0.22.0
requests==2.18.4
requests-oauthlib==0.8.0
texttable==1.2.1
beautifulsoup4==4.5.3